Add files via upload
This is the todo app, in which we can add the task, delete it or cross it according to our need. This is made using HTML, CSS and JS(Event Listener). 
I have placed event listener to the buttons such as on "+" button of task, "✔️" button on previous added task to cross the task if neeeded and "🗑️"(Trash Ircon) in app to permanently delete the task.

I used many other key concepts of JS, such as:
i) document.getElementById() --> to get the element by their id and apply JS on them.
ii)  EventListener() --> It lates the button to play roles what i want, mainly of them are:
e.target() -> to target the element
appendChild() -> to append the child element
createElement() -> to create the element, such as list for the task those are being added.
parentElement() -> to apply event listener to the parent element of some element.
remove() -> removes the element on which it is applied (used to delete the task)
classlist.add() ->used to add the element and give them a particular class.

These were some key concepts which i used in JS.